Move alias migration test to separate file
